+++ libs/fn/src/lib/cdk/selectable-list/selectable-item.token.ts
@@ -1,5 +1,5 @@
 import { Observable } from 'rxjs';
-import { HasElementRef } from '../HasElementRef';
+import { HasElementRef } from '../has-element-ref';
 
 export abstract class SelectableItemToken<ValueType = any> extends HasElementRef {
     abstract value: ValueType;
